[MIPS64] libc/libm support
libc/libm support for MIPS64 targets

+++ libc/include/sys/stat.h
@@ -62,24 +62,24 @@ struct stat {
 };
 #elif defined(__mips__)
 struct stat {
-  unsigned long st_dev;
-  unsigned long __pad0[3];
+  unsigned int st_dev;
+  unsigned int __pad0[3];
   unsigned long long st_ino;
   unsigned int st_mode;
   unsigned int st_nlink;
-  unsigned long st_uid;
-  unsigned long st_gid;
-  unsigned long st_rdev;
-  unsigned long __pad1[3];
+  unsigned int st_uid;
+  unsigned int st_gid;
+  unsigned int st_rdev;
+  unsigned int __pad1[3];
   long long st_size;
-  unsigned long st_atime;
-  unsigned long st_atime_nsec;
-  unsigned long st_mtime;
-  unsigned long st_mtime_nsec;
-  unsigned long st_ctime;
-  unsigned long st_ctime_nsec;
-  unsigned long st_blksize;
-  unsigned long __pad2;
+  unsigned int st_atime;
+  unsigned int st_atime_nsec;
+  unsigned int st_mtime;
+  unsigned int st_mtime_nsec;
+  unsigned int st_ctime;
+  unsigned int st_ctime_nsec;
+  unsigned int st_blksize;
+  unsigned int __pad2;
   unsigned long long st_blocks;
 };
 #elif defined(__x86_64__)
